document and test reduct_base part
